requirements.txt
numpy
pandas
scikit-learn
joblib
pyyaml
xgboost
mlflow
matplotlib
fastapi
uvicorn[standard]
gunicorn
